refactor: Use RestrictedViewSize in ProjectSetRestrictedViewSizeRequest, replace String
