+++ webapi/src/main/scala/org/knora/webapi/slice/admin/domain/model/RestrictedViewSize.scala
@@ -1,21 +1,18 @@
-/*
- * Copyright © 2021 - 2024 Swiss National Data and Service Center for the Humanities and/or DaSCH Service Platform contributors.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package dsp.valueobjects
-
-import zio.json.JsonCodec
+package org.knora.webapi.slice.admin.domain.model
 
 import scala.util.matching.Regex
+
+import org.knora.webapi.slice.common.StringValueCompanion
+import org.knora.webapi.slice.common.Value.StringValue
 
 /**
  * RestrictedViewSize value object.
  */
-sealed abstract case class RestrictedViewSize private (value: String)
+case class RestrictedViewSize private (value: String) extends StringValue
 
-object RestrictedViewSize {
-  def make(value: String): Either[String, RestrictedViewSize] = {
+object RestrictedViewSize extends StringValueCompanion[RestrictedViewSize] {
+  def default: RestrictedViewSize = new RestrictedViewSize("!512,512") {}
+  def from(value: String): Either[String, RestrictedViewSize] = {
     val trimmed: String = value.trim
     // matches strings "pct:1-100"
     val percentagePattern: Regex = "pct:[1-9][0-9]?0?$".r
@@ -32,13 +29,6 @@ object RestrictedViewSize {
     else Left(ErrorMessages.RestrictedViewSizeInvalid(value))
   }
 
-  def unsafeFrom(value: String): RestrictedViewSize =
-    make(value).fold(s => throw new IllegalArgumentException(s), identity)
-
-  def default: RestrictedViewSize = new RestrictedViewSize("!512,512") {}
-
-  implicit val codec: JsonCodec[RestrictedViewSize] =
-    JsonCodec[String].transformOrFail(RestrictedViewSize.make, _.value)
 }
 
 object ErrorMessages {
