+++ webapi/src/test/scala/dsp/valueobjects/RestrictedViewSizeSpec.scala
@@ -8,8 +8,10 @@ package dsp.valueobjects
 import zio.Scope
 import zio.test.*
 
+import org.knora.webapi.slice.admin.domain.model.RestrictedViewSize
+
 /**
- * This spec is used to test the [[dsp.valueobjects.RestrictedViewSize]] value object creation.
+ * This spec is used to test the [[RestrictedViewSize]] value object creation.
  */
 object RestrictedViewSizeSpec extends ZIOSpecDefault {
 
@@ -18,31 +20,31 @@ object RestrictedViewSizeSpec extends ZIOSpecDefault {
       val gen = Gen.int(1, 100)
       check(gen) { int =>
         val param = s"pct:$int"
-        assertTrue(RestrictedViewSize.make(param).map(_.value) == Right(param))
+        assertTrue(RestrictedViewSize.from(param).map(_.value) == Right(param))
       }
     },
     test("should succeed on passing the same x y dimensions") {
       val gen = Gen.int(1, 1000)
       check(gen) { integer =>
         val param = s"!$integer,$integer"
-        assertTrue(RestrictedViewSize.make(param).map(_.value) == Right(param))
+        assertTrue(RestrictedViewSize.from(param).map(_.value) == Right(param))
       }
     },
     test("should fail on passing negative x y dimensions") {
       val gen = Gen.int(-1000, -1)
       check(gen) { integer =>
         val param = s"!$integer,$integer"
-        assertTrue(RestrictedViewSize.make(param).map(_.value) == Left(ErrorMessages.RestrictedViewSizeInvalid(param)))
+        assertTrue(RestrictedViewSize.from(param).map(_.value) == Left(s"Invalid RestrictedViewSize: $param"))
       }
     },
     test("should fail on passing incorrect values") {
       val gen = Gen.fromIterable(Seq("!512,100", "pct:-1", "pct:0", "pct:101"))
       check(gen) { param =>
-        assertTrue(RestrictedViewSize.make(param) == Left(ErrorMessages.RestrictedViewSizeInvalid(param)))
+        assertTrue(RestrictedViewSize.from(param) == Left(s"Invalid RestrictedViewSize: $param"))
       }
     },
     test("should fail on passing empty value") {
-      assertTrue(RestrictedViewSize.make("") == Left(ErrorMessages.RestrictedViewSizeMissing))
+      assertTrue(RestrictedViewSize.from("") == Left("RestrictedViewSize cannot be empty."))
     }
   )
 }
